requirements.txt
dotenv-python==0.0.1
kuzu==0.6.0
neo4j==5.25.0
numpy==2.1.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
